introduce yarnclientRetriever interface
