+++ flink-yarn/src/main/java/org/apache/flink/yarn/YarnClientRetriever.java
@@ -7,7 +7,7 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -20,57 +20,10 @@ package org.apache.flink.yarn;
 
 import org.apache.flink.annotation.Internal;
 
-import org.apache.hadoop.service.Service;
 import org.apache.hadoop.yarn.client.api.YarnClient;
-import org.apache.hadoop.yarn.conf.YarnConfiguration;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import static org.apache.hadoop.yarn.client.api.YarnClient.createYarnClient;
-
-/**
- * Retriever is used to get yarn client for {@link ApplicationReportProviderImpl}. When external
- * yarn client is closed, it will create new yarn client that its lifecycle will be managed in this
- * retriever.
- */
+/** Factory for getting {@link org.apache.hadoop.yarn.client.api.YarnClient}. */
 @Internal
-public final class YarnClientRetriever implements AutoCloseable {
-    private static final Logger LOG = LoggerFactory.getLogger(YarnClientRetriever.class);
-
-    private final YarnClient externalYarnClient;
-    private final YarnConfiguration yarnConfiguration;
-    private YarnClient createdYarnClient;
-
-    private YarnClientRetriever(YarnClient yarnClient, YarnConfiguration yarnConfiguration) {
-        this.externalYarnClient = yarnClient;
-        this.yarnConfiguration = yarnConfiguration;
-    }
-
-    public YarnClient getYarnClient() {
-        if (externalYarnClient != null && !externalYarnClient.isInState(Service.STATE.STOPPED)) {
-            return externalYarnClient;
-        }
-
-        if (createdYarnClient != null) {
-            return createdYarnClient;
-        }
-
-        LOG.info("The external yarn client is closed, new yarn client will be created.");
-        this.createdYarnClient = createYarnClient();
-        createdYarnClient.init(yarnConfiguration);
-        createdYarnClient.start();
-        return createdYarnClient;
-    }
-
-    @Override
-    public void close() throws Exception {
-        if (createdYarnClient != null) {
-            createdYarnClient.stop();
-        }
-    }
-
-    public static YarnClientRetriever from(
-            YarnClient yarnClient, YarnConfiguration yarnConfiguration) {
-        return new YarnClientRetriever(yarnClient, yarnConfiguration);
-    }
+public interface YarnClientRetriever extends AutoCloseable {
+    YarnClient getYarnClient();
 }
